requirements.txt
Django==4.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pyotp==2.9.0
django-redis==5.2.0
django_environ==0.9.0
